requirements.txt
Django==3.2.18
djangorestframework==3.12.4
requests==2.28.2
beautifulsoup4==4.9.3
nltk==3.8.1
rake-nltk==1.0.6
openai==0.27.2
scikit-learn==1.2.2
numpy==1.24.2
django-cors-headers==3.14.0
spacy==3.5.2
celery==5.2.7
redis
django-celery-results==2.5.0
channels==4.0.0
youtube-transcript-api==0.6.1
duckduckgo_search===3.8.4
pandas
python-dotenv
langchain
boto3
unstructured==0.7.12
unstructured-inference
pytesseract
psycopg2-binary
typesense==0.19.0
djangorestframework-simplejwt==5.3.1